requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
frozenlist==1.4.1
geoip2==4.8.0
idna==3.6
maxminddb==2.5.2
multidict==6.0.5
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4